python/requirements.txt
grpcio=1.32.0
protobuf
pip install grpcio-status